LayoutTests/http/tests/media/audio-load-loadeddata-expected.txt


RUN(video.muted = true)
RUN(video.src = '/media/resources/load-video.py?type=audio/mpeg&name=sound_5.mp3')
RUN(video.load())
EVENT(loadeddata)
END OF TEST
